fix: many fixes and refactors

+++ app/Http/Controllers/Admin/SiteSettingController.php
@@ -56,8 +56,8 @@ class SiteSettingController extends Controller
 
             $validator = $request->validated();
 
-            $languages = $request->get('languages');
-            $slug = Str::slug($languages[2]['title']);
+            $languages = $request->input('languages', []);
+            $slug = $this->resolveSlugFromLanguages($languages);
 
             $siteSetting = $this->siteSettingRepo->create(
                 [
@@ -85,7 +85,7 @@ class SiteSettingController extends Controller
             'themes' => $this->themeRepo->toSelectTheme(),
             'languages' => Language::get(),
             'translationFields' => SiteSettingLang::translationFields(),
-            'translationValues' => []
+            'translationValues' => SiteSettingLang::translationFieldValues($siteSetting->langs)
         ]);
     }
 
@@ -96,11 +96,12 @@ class SiteSettingController extends Controller
 
             $validator = $request->validated();
 
-            $languages = $request->get('languages');
-            $slug = Str::slug($languages[2]['title']);
+            $languages = $request->input('languages', []);
+            $slug = $this->resolveSlugFromLanguages($languages);
             
             $this->siteSettingRepo->update($siteSetting->id, [
-                'slug' => $slug
+                'slug' => $slug,
+                'theme_id' => $request->input('theme_id'),
             ]);
             
             $this->siteSettingLangStrategy->decideCreateOrUpdate($request->get('languages'), $siteSetting);
@@ -117,6 +118,12 @@ class SiteSettingController extends Controller
 
     public function destroy(SiteSetting $siteSetting)
     {
+        if ($this->isProtectedSetting($siteSetting)) {
+            return redirect()
+                ->route('site.setting.edit', $siteSetting)
+                ->with('warning', __('This setting cannot be deleted.'));
+        }
+
         try {
             DB::beginTransaction();
             $siteSetting->langs()->delete();
@@ -128,5 +135,29 @@ class SiteSettingController extends Controller
         }
 
         return Inertia::location(route('site.setting.index'));
+    }
+
+    private function resolveSlugFromLanguages(array $languages): string
+    {
+        $defaultLanguageId = Language::query()->where('default', 1)->value('id');
+        $defaultTitle = $defaultLanguageId ? data_get($languages, $defaultLanguageId . '.title') : null;
+
+        $fallbackTitle = collect($languages)
+            ->pluck('title')
+            ->filter(fn ($value) => filled($value))
+            ->first();
+
+        $title = $defaultTitle ?: $fallbackTitle;
+
+        if (blank($title)) {
+            $title = Str::lower(Str::random(10));
+        }
+
+        return Str::slug($title);
+    }
+
+    private function isProtectedSetting(SiteSetting $siteSetting): bool
+    {
+        return (int) $siteSetting->id === 1;
     }
 }
